+++ src/classes/Gantry/Component/Assignments/AssignmentFilter.php
@@ -38,7 +38,7 @@ class AssignmentFilter
         }
 
         ksort($scores, SORT_STRING);
-        arsort($scores);
+        arsort($scores, SORT_NUMERIC);
 
         return $scores;
     }
@@ -103,13 +103,13 @@ class AssignmentFilter
             $this->method = 'calcMax';
         }
 
-        return $this->calcArray(null, $matches);
+        return $this->calcArray(0, $matches);
     }
 
     /**
-     * @param int|float $carry
-     * @param int|float|array $item
-     * @return int|float
+     * @param float $carry
+     * @param float|array $item
+     * @return float
      * @internal
      */
     protected function calcArray($carry, $item)
@@ -119,35 +119,35 @@ class AssignmentFilter
         }
 
         $method = $this->method;
-        return $this->{$method}($carry, $item);
+        return $this->{$method}($carry, (float) $item);
     }
 
     /**
-     * @param int|float $carry
-     * @param int|float $item
-     * @return int|float
+     * @param float $carry
+     * @param float $item
+     * @return float
      * @internal
      */
     protected function calcOr($carry, $item)
     {
-        return (int) ($carry || $item);
+        return (float) ($carry || $item);
     }
 
     /**
-     * @param int|float $carry
-     * @param int|float $item
-     * @return int|float
+     * @param float $carry
+     * @param float $item
+     * @return float
      * @internal
      */
     protected function calcMin($carry, $item)
     {
-        return isset($carry) ? min($carry, $item) : $item;
+        return $carry ? min($carry, $item) : $item;
     }
 
     /**
-     * @param int|float $carry
-     * @param int|float $item
-     * @return int|float
+     * @param float $carry
+     * @param float $item
+     * @return float
      * @internal
      */
     protected function calcMax($carry, $item)
@@ -156,9 +156,9 @@ class AssignmentFilter
     }
 
     /**
-     * @param int|float $carry
-     * @param int|float $item
-     * @return int|float
+     * @param float $carry
+     * @param float $item
+     * @return float
      * @internal
      */
     protected function calcSum($carry, $item)
@@ -167,13 +167,13 @@ class AssignmentFilter
     }
 
     /**
-     * @param int|float $carry
-     * @param int|float $item
-     * @return int|float
+     * @param float $carry
+     * @param float $item
+     * @return float
      * @internal
      */
     protected function calcMul($carry, $item)
     {
-        return isset($carry) ? $carry * $item : $item;
+        return $carry ? $carry * $item : $item;
     }
 }